pyproject.toml
[project]
name = "codevet"
version = "0.1.0"
description = "AI Code Vet / Auto-Fixer - Local CLI that vets AI-generated code in a Docker sandbox with Ollama-powered test generation and auto-fixing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Steve Solun"}
]
keywords = ["ai", "code-review", "testing", "docker", "ollama", "auto-fix"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "typer>=0.12",
    "docker>=7.0",
    "pytest>=8.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "ollama>=0.4",
]

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0",
]
dev = [
    "mypy>=1.10",
    "ruff>=0.4",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pyyaml>=6.0",
    "types-pyyaml>=6.0",
]

[project.scripts]
codevet = "codevet.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
